pyproject.toml
[tool.poetry]
name = "llm-project-A5"
version = "0.1.0"
description = ""
authors = ["Scrivane"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
google-genai = "^1.49.0"
openai = "^2.7.2"
groq = "^0.33.0"
gradio-client = "^1.13.3"
ollama = "^0.6.0"
mistralai = "^1.9.11"
zai-sdk = "^0.0.4.2"
transformers = "^4.57.3"
torch = "^2.9.1"
accelerate = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
